bus/hp_dio/98624: convert to anonymous namespace (nw)
